Reject port 0 to avoid Landlock wildcard
validatePorts accepted port 0, so AllowPorts=[0,443] under
allow_only emitted landlock.ConnectTCP(0) — a wildcard that
allows every port. deny_complement, in contrast, iterates 1..65535
so port 0 was silently excluded from the complement allow-set.
The two modes thus treated the same input oppositely.

Reject port 0 at the validator: validatePorts now drops it
(p >= 1) and ValidatePortRange returns an explicit error. Port 0
is not a valid TCP connect/bind destination — its only meaningful
use is bind(0) for ephemeral allocation, which Landlock does not
model — so refusing it at the boundary is the symmetric fix.

Also drop the now-redundant in-loop bounds check around
landlock.ConnectTCP — AllowSet is already filtered to
1 <= p <= 65535 by validatePorts, so the inner
'if port >= 0 && port <= 65535' guard was dead code and would
have silently let a port-0 entry through if validatePorts ever
weakened its lower bound.

Greptile P2 (both findings).

+++ internal/sandbox/linux_test.go
@@ -583,7 +583,7 @@ func TestShouldGateNetwork(t *testing.T) {
 		{
 			name:       "unrestricted mode never gates even with stale port rules",
 			mode:       NetworkUnrestricted,
-			portPolicy: PortPolicyEffective{Mode: "allow_only", AllowSet: []int{443}},
+			portPolicy: PortPolicyEffective{Mode: "allow_only", AllowSet: []uint16{443}},
 			want:       false,
 		},
 		{
@@ -595,13 +595,13 @@ func TestShouldGateNetwork(t *testing.T) {
 		{
 			name:       "outbound with allow ports gates and applies allow-set",
 			mode:       NetworkOutbound,
-			portPolicy: PortPolicyEffective{Mode: "allow_only", AllowSet: []int{443, 53}},
+			portPolicy: PortPolicyEffective{Mode: "allow_only", AllowSet: []uint16{443, 53}},
 			want:       true,
 		},
 		{
 			name:       "outbound with deny-derived allow set still gates",
 			mode:       NetworkOutbound,
-			portPolicy: PortPolicyEffective{Mode: "deny_complement", AllowSet: []int{80, 443}},
+			portPolicy: PortPolicyEffective{Mode: "deny_complement", AllowSet: []uint16{80, 443}},
 			want:       true,
 		},
 		{
